+++ src/main/java/org/opensearch/neuralsearch/processor/collapse/CollapseExecutor.java
@@ -4,7 +4,6 @@
  */
 package org.opensearch.neuralsearch.processor.collapse;
 
-import org.apache.lucene.search.grouping.CollapseTopFieldDocs;
 import org.apache.lucene.util.BytesRef;
 
 /**
@@ -22,19 +21,19 @@ public class CollapseExecutor {
      * @return The total number of documents that were collapsed during the operation
      */
     public int executeCollapse(CollapseDTO collapseDTO) {
-        boolean isKeywordCollapse = isCollapseOnKeywordField(collapseDTO);
-        CollapseStrategy collapseStrategy = isKeywordCollapse
+        Class<?> collapseFieldType = collapseDTO.getCollapseFieldType();
+        /**
+         * Ideally, the fieldType should be derived from CollapseContext but currently there is no way of providing searchContext in phase results processor.
+         * Through searchPhaseContext we can only get collapseBuilder which only contains the fieldName. To create collapseContext from builder we need QueryShardContext which is not present in phaseResultsProcessor.
+         * As per current code, if the group value is null when the document does not have collapse field then we cannot determine the field type. Therefore, we always follow Numeric strategy in that case.
+         * TODO: In future, when collapseContext is accessible then retrieve fieldType from there.
+         *
+         */
+        CollapseStrategy collapseStrategy = collapseFieldType == BytesRef.class
             ? CollapseStrategy.createKeywordStrategy()
             : CollapseStrategy.createNumericStrategy();
 
         collapseStrategy.executeCollapse(collapseDTO);
         return collapseStrategy.getTotalCollapsedDocsCount();
-    }
-
-    private boolean isCollapseOnKeywordField(CollapseDTO collapseDTO) {
-        return ((CollapseTopFieldDocs) collapseDTO.getCollapseQueryTopDocs()
-            .get(collapseDTO.getIndexOfFirstNonEmpty())
-            .getTopDocs()
-            .getFirst()).collapseValues[0] instanceof BytesRef;
     }
 }